+++ src/tooth/toothmetadata/metadata_test.go
@@ -1,4 +1,4 @@
-package metadata
+package toothmetadata
 
 import (
 	"testing"
